feat(match2): standardize FFA/BR input
